Modularize the ASTParserTask to support extension
Also added minor comment style fixes and unit test styles.
